% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/5_OutputAnalysis.R
\name{AnalyzeOutput_mLoci_Daisy}
\alias{AnalyzeOutput_mLoci_Daisy}
\title{Analyze output for mPlex-mLoci or DaisyDrive}
\usage{
AnalyzeOutput_mLoci_Daisy(readDirectory, saveDirectory, genotypes, collapse,
  numCores)
}
\arguments{
\item{readDirectory}{Directory where output was written to; should not end in path seperator}

\item{saveDirectory}{Directory to save analyzed data. Don't save to readDirectory}

\item{genotypes}{A list of each locus containing the genotypes of interest at that locus. Default is all genotypes}

\item{collapse}{A vector of each locus containing TRUE/FALSE. If TRUE, the genotypes of interest at that locus are collapsed and the output is the sum of all of them.}

\item{numCores}{How many cores to read file in with}
}
\description{
This function takes all the files in a directory and analyzes the population by
genotype of interest. It saves output by patch. The files are organzed
by time, then each genotype and total population.Data are analyzed by matching the genotypes of interest.
}
