% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/5_OutputAnalysis.R
\name{AnalyzeOutput_oLocus}
\alias{AnalyzeOutput_oLocus}
\title{Analyze output for mPlex-oLocus}
\usage{
AnalyzeOutput_oLocus(readDirectory, saveDirectory, alleles, collapse,
  numCores = 1)
}
\arguments{
\item{readDirectory}{directory where output was written to; should not end in path seperator}

\item{saveDirectory}{directory to save analyzed data. Default is readDirectory}

\item{alleles}{A list of lists that contain the genotypes of interest at each locus. Default is all genotypes}

\item{collapse}{A list of lists containing TRUE/FALSE for each locus. If TRUE, the genotypes of interest at that locus are collapsed and the output is the sum of all of them.}

\item{numCores}{How many cores to read files in with, default is 1}
}
\description{
Analyze output for mPlex-oLocus
}
\details{
This function takes all the files in a directory and analyzes the population by
genotype of interest. It saves output by patch, corresponding to the male and
female of each patch. Data are analyzed by matching the genotypes of interest.
}
