% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/4_ParametersSetup.R
\name{basicRepeatedReleases}
\alias{basicRepeatedReleases}
\title{Make List of Mosquito Releases}
\usage{
basicRepeatedReleases(
  releaseStart,
  releaseEnd,
  releaseInterval,
  genMos,
  genMosMate = NULL,
  numMos,
  minAge,
  maxAge,
  ageDist
)
}
\arguments{
\item{releaseStart}{Day releases start}

\item{releaseEnd}{Day releases end}

\item{releaseInterval}{Interval between releases}

\item{genMos}{List of genotypes for new Mosquitoes}

\item{numMos}{Integer number of Mosquitoes to create}

\item{minAge}{Integer specifying the minimum age of Mosquitoes}

\item{maxAge}{Integer specifying the maximum age of Mosquiotes}

\item{ageDist}{Distribution for ages of Mosquitoes. Must be length(maxAge-minAge+1)}

\item{genMostMate}{List of genotypes for new Mosquitoes' mates, default is NULL}
}
\value{
List of release dates and the population to be released on that day.
}
\description{
Sets up a release schedule for a single patch, returns a list to be used in
\code{\link{oneDay_Releases_Patch}}
}
\details{
This function creates a release schedule and the genotypes and ages
of mosquitoes to be released at that time. The release schedule is determined
by releaseStart, releaseEnd, and releaseInterval. The mosquitoes to be released
are determined by genMos (a list of genotypes), numMos (a vector of how many of
each genotype to release), and minAge/maxAge/ageDist. The last three determine
the age distribution of the mosquitoes to be released.
}
\examples{
# to setup for 3 patches but only release in the first with a defined release schedule:

patchReleases = replicate(n = 3,
                          expr = list(maleReleases = NULL,femaleReleases = NULL,
                          eggReleases=NULL,matedFemaleReleases = NULL),
                          simplify = FALSE)

set.seed(42)
patchReleases[[1]]$femaleReleases = basicRepeatedReleases(releaseStart = 5,
releaseEnd = 30,
releaseInterval = 5,
genMos = c("A", "B", "C"),
numMos = c(10, 20, 30),
minAge = 1,
maxAge = 10,
ageDist = rep(x = 1, times = 10-1+1)/(10-1+1)) #uniform distribution of ages

set.seed(42)
patchReleases[[1]]$maleReleases = basicRepeatedReleases(releaseStart = 50,
releaseEnd = 60,
releaseInterval = 1,
genMos = c("A", "B", "C"),
numMos = c(10, 20, 30),
minAge = 5,
maxAge = 10,
ageDist = c(0,0,0,0,0,1)) #all mosqutioes will be 10

set.seed(42)
patchReleases[[1]]$larvaeReleases = basicRepeatedReleases(releaseStart = 1,
releaseEnd = 5,
releaseInterval = 1,
genMos = c("A", "B", "C"),
numMos = c(10, 20, 30),
minAge = 5,
maxAge = 10,
ageDist = c(0,1,0,1,0,0)) #half of mosquitoes are 6, half are 8

set.seed(42)
patchReleases[[1]]$matedFemaleReleases = basicRepeatedReleases(releaseStart = 1,
releaseEnd = 5,
releaseInterval = 1,
genMos = c("A", "B", "C"),
genMosMate =c("B","B","C")
numMos = c(10, 20, 30),
minAge = 5,
maxAge = 10,
ageDist = c(0,1,0,1,0,0)) #half of mosquitoes are 6, half are 8

}
