% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/5_OutputAnalysis.R
\name{genOI_mLoci_Daisy}
\alias{genOI_mLoci_Daisy}
\title{Generate genotypes of interest for mPlex-mLoci or DaisyDrive}
\usage{
genOI_mLoci_Daisy(outputFile, genotypes, collapse)
}
\arguments{
\item{outputFile}{Name of file to output. Must end in .csv}

\item{genotypes}{A list of each locus containing the genotypes of interest at that locus. Default is all genotypes}

\item{collapse}{A vector of each locus containing TRUE/FALSE. If TRUE, the genotypes of interest at that locus are collapsed and the output is the sum of all of them.}
}
\description{
This function generates a *.csv of genotypes of interest and the grouping scheme, ie,
if one locus isn't interesting and should be considered the same. This file will be used
for data analysis.
}
