% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/5_OutputAnalysis.R
\name{ggColUtility}
\alias{ggColUtility}
\title{Utility to Imitate ggplot2 Colors}
\usage{
ggColUtility(n, alpha = 1)
}
\arguments{
\item{n}{number of colors}

\item{alpha}{transparency}
}
\description{
Sample at equally spaced intervals along the color wheel
}
