% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3_DriveReferenceSetups.R
\name{MakeReference_DaisyDrive}
\alias{MakeReference_DaisyDrive}
\title{Daisy Drive Offspring Reference}
\usage{
MakeReference_DaisyDrive(cRate, hRate, rRate, d, eta=NULL, phi=NULL,
omega=NULL, xiF=NULL, xiM=NULL, s=NULL)
}
\arguments{
\item{cRate}{Vector of cutting rates for each drive piece}

\item{hRate}{Vector of homing allele generation rates}

\item{rRate}{Vector of neutral allele generation rates}

\item{d}{Vector of background mutation rates at each locus}

\item{eta}{Named vector of mating fitness}

\item{phi}{Named vector of sex ratios at emerence}

\item{omega}{Named vector of adult mortality increase}

\item{xiF}{Named vector of female pupatory success}

\item{xiM}{Named vector of male pupatory success}

\item{s}{Named vector of genotype-dependent fertility reduction}
}
\value{
List of homing, cutting, and mendelian genotypes and rates.
}
\description{
Create a list specifying the offspring probability distribution.
}
\details{
This function creates a reference list for \code{\link{DaisyOffspring}}.
The number of drive elements is specified by the length of cRate. The first element
of each {c,h,r}Rate must be 0, because the first element behaves in a mendelian
fashion, and the rates of the final piece don't matter, because there is not
another piece to drive. cRate, hRate, and rRate must be the same length. This
function is similar to \code{\link{MakeReference_Multiplex_mLoci}}
and \code{\link{MakeReference_Multiplex_oLocus}}
}
\examples{
cRate <- c(0,0.4,0.7) # This drive targets 3 loci
hRate <- c(0, 0.2, 0.3)
rRate <- c(0, 0.006, 0.01)
d <- c(0.0001, 0.0001, 0.0001)

MakeReference_DaisyDrive(H,R,S,d)

}
