% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0_mPlexRun.R
\name{mPlex_oneRun}
\alias{mPlex_oneRun}
\title{Run one run of mPlex}
\usage{
mPlex_oneRun(
  seed,
  numThreads,
  networkParameters,
  reproductionReference,
  patchReleases,
  migrationMale,
  migrationFemale,
  migrationBatch,
  outputDirectory,
  reproductionType,
  verbose
)
}
\arguments{
\item{seed}{An integer seed for random number generator}

\item{numThreads}{An integer specifying the number of threads to parallelize over}

\item{networkParameters}{A list of simulation parameters}

\item{reproductionReference}{A list of reproduction and genotype specific parameters}

\item{patchReleases}{A list of releases}

\item{migrationMale}{A matrix specifying male migration rates}

\item{migrationFemale}{A matrix specifying female migration rates}

\item{migrationBatch}{A list specifing batch migration probabilities, rates, and sex ratios}

\item{outputDirectory}{String folder name to write output}

\item{reproductionType}{String specifying type of reproduction model}

\item{verbose}{Boolean, be chatty?}
}
\description{
R interface to C++ simulation code.
}
