% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/4_ParametersSetup.R
\name{NetworkParameters}
\alias{NetworkParameters}
\title{Network Parameters}
\usage{
NetworkParameters(
  nPatch,
  simTime,
  sampTime = 1,
  moveVar = 1000,
  tEgg = 1L,
  tLarva = 14L,
  tPupa = 1L,
  beta = 32,
  muAd = 0.123,
  dayGrowthRate = 1.096,
  AdPopEQ,
  runID = 1L
)
}
\arguments{
\item{nPatch}{number of \code{\link{Patch}}}

\item{simTime}{maximum time to run simulation}

\item{sampTime}{how often output is written. Default is 1 (i.e. every day)}

\item{moveVar}{variance of stochastic movement (not used in diffusion model of migration).
It affects the concentration of probability in the Dirchlet simplex, small
values lead to high variance and large values lead to low variance.}

\item{tEgg}{length of egg stage}

\item{tLarva}{length of larval instar stage}

\item{tPupa}{length of pupal stage}

\item{beta}{female egg batch size of wild-type}

\item{muAd}{wild-type daily adult mortality (1/muAd is average wild-type lifespan)}

\item{dayGrowthRate}{daily population growth rate (used to calculate equilibrium)}

\item{AdPopEQ}{a single number or vector of adult population size at equilibrium}

\item{runID}{begin counting runs with this set of parameters from this value}
}
\value{
List(nPatch=int, simTime=int, sampTime=int, moveVar=numeric,
runID=int, stageTime=vec int, beta=int, dayGrowthRate=numeric, AdPopEq=int vec,
genTime=numeric, genGrowthRate=numeric, mu=vec numeric,
thetaAq=vec numeric, alpha=vec numeric, Leq=vec int)
}
\description{
Generate parameters for simulation on a \code{\link{Network}}.
}
\details{
Parameters average generation time \eqn{g}, population growth rate \eqn{R_{m}},
aquatic mortality \eqn{\mu_{Aq}}, and aquatic survival \eqn{\theta_{Aq}}
are shared between patches and calculated by \code{\link{calcAverageGenerationTime}},
\code{\link{calcPopulationGrowthRate}}, \code{\link{calcLarvalStageMortalityRate}},
and \code{\link{calcAquaticStagesSurvivalProbability}}. \cr
Patch-specific parameters \eqn{\alpha} and \eqn{L_{eq}}
are calculated for each patch by \code{\link{calcDensityDependentDeathRate}}
and \code{\link{calcLarvalPopEquilibrium}}.
}
\examples{
This is an example
So is this

}
