% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/5_OutputAnalysis.R
\name{Plot_mPlex}
\alias{Plot_mPlex}
\title{Plot mPlex}
\usage{
Plot_mPlex(directory, whichPatches=NULL,totalPop=TRUE)
}
\arguments{
\item{directory}{Path to directory of analyzed output files}

\item{whichPatches}{Vector of patches to plot, must be less than 15}

\item{totalPop}{Boolean, to plot the total population or not.}
}
\description{
Plots the analyzed output of mPlex.
}
\details{
This function plots output from the analyze function. Setting totalPop
to FALSE keeps it from plotting the total population. The function breaks down
if you plot too many patches, but it looks bad long before that.
}
