% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/5_OutputAnalysis.R
\name{plotmPlexMult}
\alias{plotmPlexMult}
\title{Plot mPlex}
\usage{
plotmPlexMult(directory, whichPatches=NULL,totalPop=TRUE)
}
\arguments{
\item{directory}{Path to directory of analyzed output files}

\item{whichPatches}{Vector of patches to plot, must be less than 15}

\item{totalPop}{Boolean, to plot the total population or not.}

\item{nonZeroGen}{Boolean, to plot genotypes that are always zero in simulation}

\item{lwd}{Double, specify the line width for plotting}

\item{alpha}{Double, specify the opacity for plotting}
}
\description{
Plots multiple analyzed outputs of mPlex. Good for repetitions.
}
\details{
This function plots output from one or more runs of the analyze function.
totPop is not currently used.
}
