% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{run_mPlex_Cpp}
\alias{run_mPlex_Cpp}
\title{Run Multi-Plex Experiment}
\usage{
run_mPlex_Cpp(seed_, networkParameters_, reproductionReference_,
  patchReleases_, migrationMale_, migrationFemale_, migrationBatch_,
  outputDirectory_, reproductionType_, verbose_)
}
\description{
Run mPlex experiments, choosing the type of inheritance desired.
}
\examples{
\dontrun{
nope
none
sorry
}


}
