% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{run_mPlex_Cpp_repetitions}
\alias{run_mPlex_Cpp_repetitions}
\title{Run Multi-Plex Experiment}
\usage{
run_mPlex_Cpp_repetitions(seed, networkParameters_, reproductionReference_,
  patchReleases_, migrationMale_, migrationFemale_, migrationBatch_,
  output_directory, reproductionType_, verbose)
}
\description{
Run multiple mPlex experiments, choosing the type of inheritance desired.
}
\examples{
\dontrun{
nope
none
sorry
}


}
