% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{run_mPlex_repetitions}
\alias{run_mPlex_repetitions}
\title{Run Multi-Plex Experiment}
\usage{
run_mPlex_repetitions()
}
\description{
Run mPlex experiments, choosing the type of inheritance desired.
}
\examples{
nope
none
sorry

}
