% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0_mPlexRun.R
\name{runMPlex}
\alias{runMPlex}
\title{Run mPlex}
\usage{
runMPlex(
  seed = 1,
  numReps = 1,
  numThreads = 1,
  networkParameters,
  reproductionReference,
  patchReleases,
  migrationMale,
  migrationFemale,
  migrationBatch,
  reproductionType,
  outputDirectory = "./",
  verbose = FALSE
)
}
\arguments{
\item{seed}{An integer seed for random number generator. Default is 1}

\item{numReps}{An integer specifying the number of repetitions. Default is 1}

\item{numThreads}{An integer specifying the number of threads to parallelize over. Default is 1}

\item{networkParameters}{A list of simulation parameters}

\item{reproductionReference}{A list of reproduction and genotype specific parameters}

\item{patchReleases}{A list of releases}

\item{migrationMale}{A matrix specifying male migration rates}

\item{migrationFemale}{A matrix specifying female migration rates}

\item{migrationBatch}{A list specifing batch migration probabilities, rates, and sex ratios}

\item{reproductionType}{String specifying type of reproduction model}

\item{outputDirectory}{String folder name to write output, default is current directory}

\item{verbose}{Boolean, be chatty? Default is FALSE}
}
\description{
R interface to C++ simulation code.
}
