% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/5_OutputAnalysis.R
\name{SimAggregation}
\alias{SimAggregation}
\title{Aggregate mPlex output}
\usage{
SimAggregation(readDirectory, writeDirectory, simTime)
}
\arguments{
\item{readDirectory}{Directory with simulation output, all male/female/patch/experiment number files.}

\item{writeDirectory}{Directory to save analyzed data. Must have aggregation key in it!}

\item{simTime}{Simulation time}
}
\description{
This function takes all the files in a directory and aggregates them based on
an "aggKey" file supplied in the write folder. This key is generated by "genOI*"
functions for the appropriate experiment. Output is organized by time and group of
interested and saved by experiment number, sex, and patch.
File structure as shown below: \cr
\itemize{
\item someTopDirectory
\itemize{
\item readDirectory
\itemize{
\item F_Run_001_Patch_000.csv
\item F_Run_002_Patch_000.csv
\item M_Run_001_Patch_000.csv
\item M_Run_002_Patch_000.csv
}
\item writeDirectory1
\itemize{
\item AggKey1.csv
}
\item writeDirectory2
\itemize{
\item AggKey2.csv
}
\item ...
}
}
}
