% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/5_OutputAnalysis.R
\name{splitOutput}
\alias{splitOutput}
\title{Split Output by Patch}
\usage{
splitOutput(directory, numCores)
}
\arguments{
\item{directory}{Directory where output was written to; must not end in path seperator}

\item{numCores}{How many cores to use for reading/writing files}
}
\value{
*.csv files for each patch
}
\description{
Split each run output into multiple files by patch.
}
