% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3_DriveReferenceSetups.R
\name{MakeReference_DaisyDrive}
\alias{MakeReference_DaisyDrive}
\title{Daisy Drive Offspring Reference}
\usage{
MakeReference_DaisyDrive(cRateM, hRateM, rRateM, dM,
cRateF=cRateM, hRateF=hRateM, rRateF=rRateM, dF=dM,
eta=NULL, phi=NULL, omega=NULL, xiF=NULL, xiM=NULL, s=NULL)
}
\arguments{
\item{cRateM}{Vector of cutting rates for each drive piece in males}

\item{hRateM}{Vector of homing allele generation rates in males}

\item{rRateM}{Vector of neutral allele generation rates in males}

\item{cRateF}{Vector of cutting rates for each drive piece in females}

\item{hRateF}{Vector of homing allele generation rates in females}

\item{rRateF}{Vector of neutral allele generation rates in females}

\item{dM}{Vector of background mutation rates at each locus in males}

\item{dF}{Vector of background mutation rates at each locus in females}

\item{eta}{Named vector of mating fitness}

\item{phi}{Named vector of sex ratios at emerence}

\item{omega}{Named vector of adult mortality increase}

\item{xiF}{Named vector of female pupatory success}

\item{xiM}{Named vector of male pupatory success}

\item{s}{Named vector of genotype-dependent fertility reduction}
}
\value{
List of homing, cutting, and mendelian genotypes and rates.
}
\description{
Create a list specifying the offspring probability distribution.
}
\details{
This function creates a reference list for \code{\link{DaisyOffspring}}.
The number of drive elements is specified by the length of cRateM. The first element
of each {c,h,r}Rate must be 0, because the first element behaves in a mendelian
fashion, and the rates of the final piece don't matter, because there is not
another piece to drive. cRateM, hRateM, and rRateM must be the same length. Each
rate can be different between males and females. The default is for females to be
the same as males, and the user only needs to specify male rates. This
function is similar to \code{\link{MakeReference_Multiplex_mLoci}}
and \code{\link{MakeReference_Multiplex_oLocus}}
}
\examples{
cRateM <- c(0,0.4,0.7) # This drive targets 3 loci
hRateM <- c(0, 0.2, 0.3)
rRateM <- c(0, 0.006, 0.01)
dM <- c(0.0001, 0.0001, 0.0001)

MakeReference_DaisyDrive(cRateM=cRateM, hRateM=hRateM, rRateM=rRateM,dM=dM)

}
