% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3_DriveReferenceSetups.R
\name{MakeReference_Multiplex_oLocus}
\alias{MakeReference_Multiplex_oLocus}
\title{mPlex One Locus Offspring Reference}
\usage{
MakeReference_Multiplex_oLocus(cRate, hRate, rRate, d, eta=NULL, phi=NULL,
omega=NULL, xiF=NULL, xiM=NULL, s=NULL)
}
\arguments{
\item{cRate}{Vector of cutting rates for each drive piece}

\item{hRate}{Vector of homing allele generation rates}

\item{rRate}{Vector of neutral allele generation rates}

\item{d}{Vector of background mutation rates at each locus}

\item{eta}{Named vector of mating fitness}

\item{phi}{Named vector of sex ratios at emerence}

\item{omega}{Named vector of adult mortality increase}

\item{xiF}{Named vector of female pupatory success}

\item{xiM}{Named vector of male pupatory success}

\item{s}{Named vector of genotype-dependent fertility reduction}
}
\value{
List of homing and mendelian genotypes and rates.
}
\description{
Create a list specifying the offspring probability distribution.
}
\details{
This function creates a reference list for \code{\link{MultiplexOffspring_oLocus}}.
It assumes multiple targeting gRNAs for 1 locus, all targets segregate together.
The length of each allele is specified by the length of H. R, S, and d must
be the same length as H, but will generally be the same number replicated that
many times.
This function is similar to \code{\link{MakeReference_DaisyDrive}} and
\code{\link{MakeReference_Multiplex_mLoci}}
}
\examples{
cRate <- c(0.9,0.4,0.7) # This drive targets 3 loci
hRate <- c(0.1, 0.2, 0.3)
rRate <- c(0.003, 0.006, 0.01)
d <- c(0.0001, 0.0001, 0.0001)

MakeReference_Multiplex_oLocus(H,R,S,d)

}
