% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0_mPlexRun.R
\name{mPlex_runRepetitions}
\alias{mPlex_runRepetitions}
\title{Run multiple runs of mPlex}
\usage{
mPlex_runRepetitions(seed, networkParameters, reproductionReference,
  patchReleases, migrationMale, migrationFemale, migrationBatch,
  output_directory, reproductionType, verbose)
}
\arguments{
\item{seed}{An integer seed for random number generator}

\item{networkParameters}{A list of simulation parameters}

\item{reproductionReference}{A list of reproduction and genotype specific parameters}

\item{patchReleases}{A list of releases}

\item{migrationMale}{A matrix specifying male migration rates}

\item{migrationFemale}{A matrix specifying female migration rates}

\item{migrationBatch}{A list specifing batch migration probabilities, rates, and sex ratios}

\item{output_directory}{Vector of string folder name to write output}

\item{reproductionType}{String specifying type of reproduction model}

\item{verbose}{Boolean, be chatty?}
}
\description{
R interface to C++ simulation code.
}
