% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/5_OutputAnalysis.R
\name{splitOutput}
\alias{splitOutput}
\title{Split Output by Patch}
\usage{
splitOutput(readDir, writeDir, remFile, numCores)
}
\arguments{
\item{readDirectory}{Directory where output was written to}

\item{saveDirectory}{Directory to write output to. Default is readDir}

\item{remFile}{Remove original output? Default is TRUE}

\item{numCores}{How many cores to use when writing output. Default is 1}
}
\value{
*.csv files for each patch
}
\description{
Split each run output into multiple files by patch.
}
